tests: fix signed vs unsigned warnings
Building the testsuite yields a bunch of signed vs unsigned warnings,
fix all of them.

+++ tests/get_name.c
@@ -52,7 +52,7 @@ static void check_name(void *fdt, const char *path)
 		FAIL("fdt_get_name(%s) returned \"%s\" instead of \"%s\"",
 		     path, getname, checkname);
 
-	if (len != strlen(getname))
+	if ((size_t)len != strlen(getname))
 		FAIL("fdt_get_name(%s) returned length %d instead of %zd",
 		     path, len, strlen(getname));
